fix: preserve size report helper on base checkout
