+++ scripts/size-report-package.mjs
@@ -7,10 +7,20 @@ import {
   formatMaybeBytes,
   formatSignedBytes,
 } from './size-report-format.mjs';
-import {
-  assertSnapshotBridgeAssets,
-  SNAPSHOT_BRIDGE_ASSET_PATHS,
-} from './lib/snapshot-bridge-assets.mjs';
+
+export const SNAPSHOT_BRIDGE_ASSET_PATHS = Object.freeze([
+  'apple/snapshot-bridge/SnapshotBridge.m',
+  'apple/snapshot-bridge/SnapshotBridgeRuntime.m',
+  'apple/snapshot-bridge/SnapshotBridgeRuntime.h',
+]);
+
+export function assertSnapshotBridgeAssets(presentPaths, context) {
+  const present = new Set(presentPaths);
+  const missing = SNAPSHOT_BRIDGE_ASSET_PATHS.filter((assetPath) => !present.has(assetPath));
+  if (missing.length > 0) {
+    throw new Error(`${context} is missing: ${missing.join(', ')}`);
+  }
+}
 
 const PACKAGE_COMPONENTS = [
   {
